feat: Extract isSameOrigin utility function
