+++ app/_features/viewTransition/index.ts
@@ -1,1 +1,2 @@
 export * from "./UniversalLink";
+export * from "./isSameOrigin";
